detect if a pid exists before signaling it
The code currently tries to detect if a pid exists and is able to be
signaled before signaling it, however os.FindProcess() always succeeds
on Unix. This change switches to using signal(0) to determine if the pid
is live instead (as documented in the kill syscall's man page).

Previously, for a nonexistent pid:
  $ ./main
  2023/08/08 16:44:05 seamless: Starting child process
  2023/08/08 16:44:06 seamless: Notifying old process
  2023/08/08 16:44:06 seamless: Could not send SIGTERM to old process: os: process already finished

Now, for a nonexistent pid:
  $ ./main
  2023/08/08 16:41:55 seamless: Starting child process
  2023/08/08 16:41:56 seamless: Notifying old process
  2023/08/08 16:41:56 seamless: Could not find old process: os: process already finished